+++ include/Word/ale_word.h
@@ -90,7 +90,7 @@ struct ALEWord {
     char     address[4];     ///< 3 decoded characters + null terminator
     uint32_t raw_payload;    ///< Raw 21-bit payload
     uint8_t  fec_errors;     ///< Golay errors corrected
-    uint32_t unanimous_votes;///< 2/3-voter unanimous count 0..48 (A.5.2.6.3)
+    uint8_t  unanimous_votes;///< 2/3-voter unanimous count 0..48 (A.5.2.6.3)
     bool     valid;          ///< Word passed FEC and character validation
     uint32_t timestamp_ms;   ///< Reception timestamp (ms)
